Move GraphContext to OrdinoAppComponent
