+++ src/internal/components/dataset/NamedSetList.tsx
@@ -3,8 +3,8 @@ import React from 'react';
 import {Button, ButtonGroup, Col, Dropdown} from 'react-bootstrap';
 import {INamedSet} from 'tdp_core';
 import {SESSION_KEY_NEW_ENTRY_POINT} from '../..';
-import {GraphContext} from '../../menu/StartMenuReact';
 import {ListItemDropdown} from '../common';
+import {GraphContext} from '../../OrdinoAppComponent';
 
 interface INamedSetListProps {
   headerIcon: string;
