+++ src/internal/components/session/TemporarySessionCard.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import {Button, Dropdown} from 'react-bootstrap';
 import {ProvenanceGraphMenuUtils} from 'tdp_core';
 import {useAsync} from '../../../hooks';
-import {GraphContext} from '../../menu/StartMenuReact';
+import {GraphContext} from '../../OrdinoAppComponent';
 import {ListItemDropdown} from '../common';
 import {CommonSessionCard} from './CommonSessionCard';
 import {SessionListItem} from './SessionListItem';
